+++ agrest-cayenne/src/main/java/io/agrest/cayenne/CayenneResolvers.java
@@ -1,10 +1,10 @@
 package io.agrest.cayenne;
 
 import io.agrest.cayenne.persister.ICayennePersister;
+import io.agrest.cayenne.processor.ICayenneQueryAssembler;
 import io.agrest.cayenne.processor.select.CayenneNestedDataResolverBuilder;
 import io.agrest.cayenne.processor.select.CayenneRootDataResolverBuilder;
 import io.agrest.runtime.AgRuntime;
-import io.agrest.runtime.path.IPathDescriptorManager;
 
 import javax.ws.rs.core.Configuration;
 
@@ -18,32 +18,32 @@ import javax.ws.rs.core.Configuration;
 public class CayenneResolvers {
 
     public static CayenneRootDataResolverBuilder root(Configuration config) {
-        return new CayenneRootDataResolverBuilder(persister(config), pathDescriptorManager(config));
+        return new CayenneRootDataResolverBuilder(persister(config), queryAssembler(config));
     }
 
     /**
      * @since 3.7
      */
-    public static CayenneRootDataResolverBuilder root(ICayennePersister persister, IPathDescriptorManager pathDescriptorManager) {
-        return new CayenneRootDataResolverBuilder(persister, pathDescriptorManager);
+    public static CayenneRootDataResolverBuilder root(ICayennePersister persister, ICayenneQueryAssembler queryAssembler) {
+        return new CayenneRootDataResolverBuilder(persister, queryAssembler);
     }
 
     public static CayenneNestedDataResolverBuilder nested(Configuration config) {
-        return new CayenneNestedDataResolverBuilder(persister(config), pathDescriptorManager(config));
+        return new CayenneNestedDataResolverBuilder(persister(config), queryAssembler(config));
     }
 
     /**
      * @since 3.7
      */
-    public static CayenneNestedDataResolverBuilder nested(ICayennePersister persister, IPathDescriptorManager pathDescriptorManager) {
-        return new CayenneNestedDataResolverBuilder(persister, pathDescriptorManager);
+    public static CayenneNestedDataResolverBuilder nested(ICayennePersister persister, ICayenneQueryAssembler queryAssembler) {
+        return new CayenneNestedDataResolverBuilder(persister, queryAssembler);
     }
 
     private static ICayennePersister persister(Configuration config) {
         return AgRuntime.service(ICayennePersister.class, config);
     }
 
-    private static IPathDescriptorManager pathDescriptorManager(Configuration config) {
-        return AgRuntime.service(IPathDescriptorManager.class, config);
+    private static ICayenneQueryAssembler queryAssembler(Configuration config) {
+        return AgRuntime.service(ICayenneQueryAssembler.class, config);
     }
 }