+++ agrest-cayenne/src/main/java/io/agrest/cayenne/processor/select/CayenneRootDataResolverBuilder.java
@@ -1,9 +1,9 @@
 package io.agrest.cayenne.processor.select;
 
 import io.agrest.cayenne.persister.ICayennePersister;
+import io.agrest.cayenne.processor.ICayenneQueryAssembler;
 import io.agrest.resolver.RootDataResolver;
 import io.agrest.resolver.RootDataResolverFactory;
-import io.agrest.runtime.path.IPathDescriptorManager;
 import org.apache.cayenne.map.ObjEntity;
 
 /**
@@ -12,11 +12,11 @@ import org.apache.cayenne.map.ObjEntity;
 public class CayenneRootDataResolverBuilder {
 
     private final ICayennePersister persister;
-    private final IPathDescriptorManager pathDescriptorManager;
+    private final ICayenneQueryAssembler queryAssembler;
 
-    public CayenneRootDataResolverBuilder(ICayennePersister persister, IPathDescriptorManager pathDescriptorManager) {
+    public CayenneRootDataResolverBuilder(ICayennePersister persister, ICayenneQueryAssembler queryAssembler) {
         this.persister = persister;
-        this.pathDescriptorManager = pathDescriptorManager;
+        this.queryAssembler = queryAssembler;
     }
 
     public RootDataResolverFactory viaQuery() {
@@ -25,9 +25,7 @@ public class CayenneRootDataResolverBuilder {
 
     protected <T> RootDataResolver<T> viaQuery(Class<?> rootType) {
         validateRoot(rootType);
-        return (RootDataResolver<T>) new ViaQueryResolver(
-                new CayenneQueryAssembler(persister.entityResolver(), pathDescriptorManager),
-                persister);
+        return (RootDataResolver<T>) new ViaQueryResolver(queryAssembler, persister);
     }
 
     protected void validateRoot(Class<?> rootType) {