+++ agrest-cayenne/src/main/java/io/agrest/cayenne/qualifier/IQualifierParser.java
@@ -1,0 +1,14 @@
+package io.agrest.cayenne.qualifier;
+
+import io.agrest.base.protocol.CayenneExp;
+import org.apache.cayenne.exp.Expression;
+
+import java.util.List;
+
+/**
+ * @since 3.3
+ */
+public interface IQualifierParser {
+
+    Expression parse(List<CayenneExp> qualifiers);
+}