+++ agrest-cayenne/src/test/java/io/agrest/cayenne/processor/CayenneQueryAssemblerTest.java
@@ -1,27 +1,18 @@
-package io.agrest.cayenne.processor.select;
+package io.agrest.cayenne.processor;
 
 import io.agrest.RootResourceEntity;
+import io.agrest.base.protocol.CayenneExp;
 import io.agrest.base.protocol.Dir;
 import io.agrest.base.protocol.Sort;
 import io.agrest.cayenne.cayenne.main.E1;
-import io.agrest.cayenne.processor.CayenneProcessor;
 import io.agrest.cayenne.unit.CayenneNoDbTest;
 import io.agrest.runtime.processor.select.SelectContext;
-import org.apache.cayenne.exp.Expression;
 import org.apache.cayenne.query.SelectQuery;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 public class CayenneQueryAssemblerTest extends CayenneNoDbTest {
-
-    private CayenneQueryAssembler queryAssembler;
-
-    @BeforeEach
-    public void before() {
-        this.queryAssembler = new CayenneQueryAssembler(runtime.getChannel().getEntityResolver(), pathDescriptorManager);
-    }
 
     @Test
     public void testCreateRootQuery_Ordering() {
@@ -78,17 +69,16 @@ public class CayenneQueryAssemblerTest extends CayenneNoDbTest {
 
     @Test
     public void testCreateRootQuery_Qualifier() {
-        Expression extraQualifier = E1.NAME.eq("X");
         RootResourceEntity<E1> entity = getResourceEntity(E1.class);
 
         SelectContext<E1> c = new SelectContext<>(E1.class);
         c.setEntity(entity);
 
-        entity.andQualifier(extraQualifier);
+        entity.getQualifiers().add(new CayenneExp("name = 'X'"));
         SelectQuery<E1> q1 = queryAssembler.createRootQuery(c);
-        assertEquals(extraQualifier, q1.getQualifier());
+        assertEquals(E1.NAME.eq("X"), q1.getQualifier());
 
-        entity.andQualifier(E1.NAME.in("a", "b"));
+        entity.getQualifiers().add(new CayenneExp("name in ('a', 'b')"));
         SelectQuery<E1> q2 = queryAssembler.createRootQuery(c);
         assertEquals(E1.NAME.eq("X").andExp(E1.NAME.in("a", "b")), q2.getQualifier());
     }