+++ agrest-sencha/src/main/java/io/agrest/sencha/runtime/processor/select/SenchaCreateResourceEntityStage.java
@@ -1,20 +1,17 @@
 package io.agrest.sencha.runtime.processor.select;
 
 import io.agrest.ResourceEntity;
+import io.agrest.base.protocol.CayenneExp;
 import io.agrest.meta.AgEntity;
-import io.agrest.runtime.entity.ICayenneExpMerger;
-import io.agrest.runtime.entity.IExcludeMerger;
-import io.agrest.runtime.entity.IIncludeMerger;
-import io.agrest.runtime.entity.IMapByMerger;
-import io.agrest.runtime.entity.ISizeMerger;
-import io.agrest.runtime.entity.ISortMerger;
+import io.agrest.runtime.entity.*;
 import io.agrest.runtime.meta.IMetadataService;
 import io.agrest.runtime.processor.select.CreateResourceEntityStage;
 import io.agrest.runtime.processor.select.SelectContext;
 import io.agrest.sencha.SenchaRequest;
 import io.agrest.sencha.runtime.entity.ISenchaFilterExpressionCompiler;
 import org.apache.cayenne.di.Inject;
-import org.apache.cayenne.exp.Expression;
+
+import java.util.List;
 
 public class SenchaCreateResourceEntityStage extends CreateResourceEntityStage {
 
@@ -41,15 +38,11 @@ public class SenchaCreateResourceEntityStage extends CreateResourceEntityStage {
     protected <T> void doExecute(SelectContext<T> context) {
         super.doExecute(context);
 
-        ResourceEntity<T> resourceEntity = context.getEntity();
-
-        Expression e1 = parseFilter(resourceEntity.getAgEntity(), context);
-        if (e1 != null) {
-            resourceEntity.andQualifier(e1);
-        }
+        ResourceEntity<T> entity = context.getEntity();
+        entity.getQualifiers().addAll(parseFilter(entity.getAgEntity(), context));
     }
 
-    protected <T> Expression parseFilter(AgEntity<?> entity, SelectContext<T> context) {
+    protected <T> List<CayenneExp> parseFilter(AgEntity<?> entity, SelectContext<T> context) {
         SenchaRequest senchaRequest = SenchaRequest.get(context);
         return senchaFilterProcessor.process(entity, senchaRequest.getFilters());
     }